+++ R/MR3d.R
@@ -7,7 +7,7 @@
 #' @importFrom rgl title3d
 #' @param seed To generate the same set of item(s) on the local computer.
 #' @param angle,x,y,z  See details
-#' @param cubes The number of connected cubes together.
+#' @param cubes The number of connected cubes to generate.
 #' @param axis Showing the axis is helpful when first testing the function.
 #' @param filled yes, no, random will determine the colour of the cubes.
 #' @description This function generates a 3 dimensional display figure. It acts as a wrap because the creation of the figure is done using functions from the rgl package.
@@ -31,11 +31,11 @@
 #' \item{rotationMatrix}{Return the cordinates of x,y,z,w}
 #'     }
 #' @examples
-#' item <- spatial3d(seed=4, angle=pi/1.3, x=0.3,y=4,z=0.8,axis = TRUE)
+#' item <- spatial3d(seed=4, angle=pi/1.3, x=0.3,y=4,z=0.8,cubes=8,axis = TRUE)
 #'
 #' # To save the figure (not run)
 #' # library(rgl)
-#' # item <- spatial3d(seed=4, angle=pi/1.3, x=0.3,y=4,z=0.8,cubes=9,axis = TRUE)
+#' # item <- spatial3d(seed=4, angle=pi/1.3, x=0.3,y=4,z=0.8,cubes=8,axis = TRUE)
 #'
 #' # save in pdf
 #' # wd<- '~/desktop'
@@ -48,7 +48,7 @@
 #' # rgl.snapshot(filename="image3D.png",fmt="png")
 
 spatial3d <- function(seed=1, angle=pi/1.3, x=0.3,y=3, z=0.8, cubes=8, axis = TRUE, filled='yes'){
-  #seed=1; angle=pi/1.3; x=0.3;y=3; z=0.8; cubes=8; axis = TRUE; filled='random'
+  #seed=1; angle=pi/1.3; x=0.3;y=3; z=0.8; cubes=8; axis = TRUE; filled='yes'
 
   # This will finalise the item
   set.seed(seed)
@@ -77,16 +77,14 @@ spatial3d <- function(seed=1, angle=pi/1.3, x=0.3,y=3, z=0.8, cubes=8, axis = TR
       cube(vlist[1],vlist[2],vlist[3],filled=FALSE)
     }
 
-
     (step <- sample(1:3, 1))
     (vlist[step] <- vlist[step]+1^rbinom(1,1,.25))
     (zlist[i,] <- vlist)
-    zlist
   }
 
-  zlist
-  #clear3d()
-  #cube(zlist[1:cubes,1],zlist[1:cubes,2],zlist[1:cubes,3])
+
+  clear3d()
+  cube(zlist[1:cubes,1],zlist[1:cubes,2],zlist[1:cubes,3]) # this is the final image to be mirrored
   rgl.viewpoint(fov = 0, userMatrix = rotationMatrix(angle=angle, x=x,y=y,z=z))
   roMatrix <- rotationMatrix(angle=angle, x=x,y=y,z=z)
   if(axis == TRUE){
@@ -106,7 +104,8 @@ spatial3d <- function(seed=1, angle=pi/1.3, x=0.3,y=3, z=0.8, cubes=8, axis = TR
 #' @importFrom rgl axes3d
 #' @importFrom rgl title3d
 #' @param obj An object with class of threeD.
-#' @param angle,x,y,z  See details
+#' @param angle,x,y,z  See details.
+#' @param cubes The number of connected cubes to generate.
 #' @param axis Showing the axis is helpful when first testing the function.
 #' @description This function generates the mirror image of the 3 dimensional display figure. It acts as a wrap because the creation of the figure is done using functions from the rgl package.
 #' @details For 3D figures, some of the cubes may be hidden in sight when automatically generated. Hence, one would need to rotate the display figure several times to ensure that none of the cubes are hidden.
@@ -128,33 +127,38 @@ spatial3d <- function(seed=1, angle=pi/1.3, x=0.3,y=3, z=0.8, cubes=8, axis = TR
 #' \item{figure}{Return the matrix that generates the mirror image of the display figure.}
 #'     }
 #' @examples
-#' display <- spatial3d(seed=4, angle=pi/1.3, x=0.3,y=4,z=0.8,axis = TRUE)
-#' display_mirror <- spatial3d_mirror(display, angle=pi/1.3, x=0.3,y=4,z=0.8,axis = TRUE)
+#' display <- spatial3d(seed=4, angle=pi/1.3, x=0.3,y=4,z=0.8, cubes=9, axis = TRUE)
+#' display_mirror <- spatial3d_mirror(display, angle=pi/1.3, x=0.3,y=4,z=0.8, cubes=9, axis = TRUE)
 #'
 #' # To save the figure (not run)
 #' # library(rgl)
-#' # display_mirror <- spatial3d_mirror(display, angle=pi/1.3, x=0.3,y=4,z=0.8,axis = TRUE)
+#' # display_mirror <- spatial3d_mirror(display, angle=pi/1.3, x=0.3,y=4,z=0.8, cubes=9, axis = TRUE)
 #' # wd<- '~/desktop'
 #'
 #' # save in pdf
 #' # item <- 1
-#' # save <- paste0(wd,"/mirror3d_",item,".pdf")
 #' # rgl.postscript(save,"pdf")
 #'
 #' # save in png
 #' # rgl.snapshot(filename="image3D.png",fmt="png")
 #'
 
-spatial3d_mirror <- function(obj, angle=pi/1.3, x=0.3,y=3, z=0.8,axis = TRUE){
+spatial3d_mirror <- function(obj, angle=pi/1.3, x=0.3,y=3, z=0.8,cubes=1,axis = TRUE){
   result <- obj
+
+  if(cubes > nrow(result$figure)){
+    stop("You need to reduce the number of cubes.")
+  }
   if(class(result)!="threeD"){
     stop("You need to use an object of class threeD.")
   }
+
   clear3d()
   rgl.light()
-  cube(result$figure[1:nrow(result$figure),3],result$figure[1:nrow(result$figure),2],result$figure[1:nrow(result$figure),1]) #mirror
+  cube(result$figure[1:cubes,3],result$figure[1:cubes,2],result$figure[1:cubes,1]) #mirror
   rgl.viewpoint(fov = 0, userMatrix = rotationMatrix(angle=angle, x=x,y=y,z=z))
-  mirrorResult<- cbind(result$figure[1:nrow(display$figure),3],result$figure[1:nrow(display$figure),2],result$figure[1:nrow(display$figure),1])
+  mirrorResult<- cbind(result$figure[1:cubes,3],result$figure[1:cubes,2],result$figure[1:cubes,1])
+
   if(axis == TRUE){
     axes3d( edges=c("x", "y", "z") )
     title3d(main=NULL, sub=NULL, xlab='xlab', ylab='ylab', zlab='zlab')